backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ollama==0.1.7
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
chromadb==0.4.22
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
torch==2.1.1
transformers==4.36.2
openai-whisper==20231117
ftfy==6.1.3
regex==2023.10.3
Pillow==10.1.0
torchvision==0.16.1 